upgrade to using lire1.0b2. all LIRE features supported.

+++ src/main/java/org/elasticsearch/index/query/image/AbstractImageScorer.java
@@ -1,6 +1,5 @@
 package org.elasticsearch.index.query.image;
 
-import net.semanticmetadata.lire.imageanalysis.LireFeature;
 import org.apache.lucene.index.BinaryDocValues;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.LeafReader;
@@ -8,6 +7,8 @@ import org.apache.lucene.search.Scorer;
 import org.apache.lucene.search.Weight;
 import org.apache.lucene.util.BytesRef;
 import org.elasticsearch.ElasticsearchImageProcessException;
+
+import net.semanticmetadata.lire.imageanalysis.features.LireFeature;
 
 import java.io.IOException;
 
@@ -45,14 +46,14 @@ public abstract class AbstractImageScorer extends Scorer {
             LireFeature docFeature = lireFeature.getClass().newInstance();
             docFeature.setByteArrayRepresentation(bytesRef.bytes);
 
-            float distance = lireFeature.getDistance(docFeature);
-            float score;
-            if (Float.compare(distance, 1.0f) <= 0) { // distance less than 1, consider as same image
+            double distance = lireFeature.getDistance(docFeature);
+            double score;
+            if (Double.compare(distance, 1.0f) <= 0) { // distance less than 1, consider as same image
                 score = 2f - distance;
             } else {
                 score = 1 / distance;
             }
-            return score * boost;
+            return (float)score * boost;
         } catch (Exception e) {
             throw new ElasticsearchImageProcessException("Failed to calculate score", e);
         }
